16nov2013 : Synchronisation NotePicker / DataView

+++ BerichtsheftApp/src/com/applang/berichtsheft/BerichtsheftApp.java
@@ -7,6 +7,8 @@ import java.io.Reader;
 import java.io.StringReader;
 import java.util.Properties;
 
+import javax.xml.transform.ErrorListener;
+import javax.xml.transform.TransformerException;
 import javax.xml.transform.TransformerFactory;
 import javax.xml.transform.sax.SAXResult;
 import javax.xml.transform.sax.SAXSource;
@@ -17,6 +19,7 @@ import javax.xml.transform.stream.StreamSource;
 import org.apache.xml.serializer.OutputPropertiesFactory;
 import org.apache.xml.serializer.Serializer;
 import org.apache.xml.serializer.SerializerFactory;
+import org.gjt.sp.jedit.View;
 import org.gjt.sp.jedit.jEdit;
 import org.xml.sax.InputSource;
 import org.xml.sax.XMLReader;
@@ -40,11 +43,6 @@ public class BerichtsheftApp
 	public static void loadSettings() {
 		System.setProperty("settings.dir", ".jedit/plugins/berichtsheft");
 		Settings.load();
-	}
-	
-	public static String berichtsheftPath(String...parts) {
-		parts = arrayappend(strings(System.getProperty("settings.dir", "")), parts);
-		return pathCombine(parts);
 	}
 	/**
 	 * @param args
@@ -87,8 +85,13 @@ public class BerichtsheftApp
 		return activity;
 	}
 	
-	public static org.gjt.sp.jedit.View getJEditView() {
-		return (org.gjt.sp.jedit.View)Activity.frame;
+	public static View getJEditView() {
+		return (View)Activity.frame;
+	}
+	
+	public static String berichtsheftPath(String...parts) {
+		parts = arrayappend(strings(System.getProperty("settings.dir", "")), parts);
+		return pathCombine(parts);
 	}
 	
 	public static String prompt(int type, String title, String message, String[] values, String...defaults) {
@@ -111,20 +114,23 @@ public class BerichtsheftApp
 	}
 	
 	public static String parameters(Object... params) {
+		String[] dbFiles = param(strings(), 0, params);
 		return String.format(
 			"<params>" +
 				"<dbfile>%s</dbfile>" +
+				"<dbfile2>%s</dbfile2>" +
 				"<year>%d</year>" +
 				"<weekInYear>%d</weekInYear>" +
 				"<dayInWeek>%s</dayInWeek>" +
 			"</params>",
-			paramString("", 0, params),
-			paramInteger(2013, 1, params),
-			paramInteger(1, 2, params),
-			paramString("\\d", 3, params));
+			param("", 0, dbFiles),
+			param("", 1, dbFiles),
+			param_Integer(2013, 1, params),
+			param_Integer(1, 2, params),
+			param_String("\\d", 3, params));
 	}
 
-	public static boolean manipContent(int phase, String vorlage, String dokument, Job<File> manipulation, Object... params) {
+	public static boolean manipContent(int phase, String vorlage, String dokument, Job<File> manipulation, Object...params) {
 		boolean begin = phase > -1;
 		boolean end = phase < 1;
 		File tempDir = tempDir(begin, BerichtsheftApp.NAME, "odt");
@@ -141,10 +147,8 @@ public class BerichtsheftApp
 						tempDir.getPath()), false);
 				archive.delete();
 			}
-			
 			if (manipulation != null)
 				manipulation.perform(new File(tempDir, "content.xml"), params);
-			
 			if (end) {
 				File destination = new File(dokument);
 				if (destination.exists())
@@ -170,46 +174,53 @@ public class BerichtsheftApp
 			return false;
 		}
 		finally {
-			if (end)
+			boolean keep = phase == 0 && param_Boolean(false, 0, params);
+			if (end && !keep)
 				deleteDirectory(tempDir);
 		}
 	}
 
-	public static boolean export(String vorlage, String dokument, final String databaseFilename, Object... params) {
-		final Integer year = paramInteger(2013, 0, params);
-		final Integer weekInYear = paramInteger(1, 1, params);
-		final String dayInWeek = paramString("\\d", 2, params);
-		
+	public static boolean export(String vorlage, String dokument, final String[] databaseFilenames, Object...params) {
+		final Integer year = param_Integer(2013, 0, params);
+		final Integer weekInYear = param_Integer(1, 1, params);
+		final String dayInWeek = param_String("\\d", 2, params);
+		boolean keep = param_Boolean(false, 3, params);
 		if (dokument.endsWith("_"))
 			dokument = dokument + String.format("%d_%d", year, weekInYear) + ".odt";
-		
 		return manipContent(
-				0, 
-				vorlage, 
-				dokument, 
-				new Job<File>() {
-					public void perform(File content, Object[] params) throws Exception {
-						File _content = new File(content.getParent(), "_content.xml");
-						content.renameTo(_content);
-						
-						File database = new File(databaseFilename);
+			0, 
+			vorlage, 
+			dokument, 
+			new Job<File>() {
+				public void perform(File content, Object[] params) throws Exception {
+					File _content = new File(content.getParent(), "_content.xml");
+					content.renameTo(_content);
+					for (int i = 0; i < databaseFilenames.length; i++) {
+						File database = new File(databaseFilenames[i]);
 						if (!database.exists())
-							throw new Exception(String.format("Database '%s' missing", database));
-						
-						String parameters = parameters(
-							databaseFilename,
-							year,
-							weekInYear,
-							dayInWeek);
-						
-						pipe(_content.getPath(), content.getPath(), new StringReader(parameters));
-						
-						_content.delete();
+							throw new Exception(
+									String.format("Database '%s' missing", database.getPath()));
 					}
-				});
+					String parameters = parameters(
+						databaseFilenames,
+						year,
+						weekInYear,
+						dayInWeek);
+					pipe(_content.getPath(), content.getPath(), new StringReader(parameters));
+					boolean keep = param_Boolean(false, 0, params);
+					if (keep) {
+						File destDir = _content.getParentFile().getParentFile();
+						copyFile(_content, new File(destDir, "_content.xml"));
+					}
+					_content.delete();
+				}
+			}, keep) && piped;
 	}
+	
+	private static boolean piped = true;
 
 	public static boolean pipe(String inputFilename, String outputFilename, Reader params) throws Exception {
+		piped = true;
 	  	TransformerFactory tFactory = TransformerFactory.newInstance();
 	  	
 	    // Determine whether the TransformerFactory supports the use of SAXSource and SAXResult
@@ -222,6 +233,22 @@ public class BerichtsheftApp
 		TransformerHandler tHandler1 = saxTFactory.newTransformerHandler(new StreamSource(getSetting("control.xsl", BerichtsheftApp.berichtsheftPath("Skripte/control.xsl"))));
 		TransformerHandler tHandler2 = saxTFactory.newTransformerHandler(new StreamSource(getSetting("content.xsl", BerichtsheftApp.berichtsheftPath("Skripte/content.xsl"))));
 		tHandler2.getTransformer().setParameter("inputfile", inputFilename);
+		tHandler2.getTransformer().setErrorListener(new ErrorListener() {
+			@Override
+			public void warning(TransformerException exception) throws TransformerException {
+				Log.w(TAG, exception.getMessage());
+			}
+			@Override
+			public void fatalError(TransformerException exception) throws TransformerException {
+				Log.e(TAG, exception.getMessage());
+				piped = false;
+			}
+			@Override
+			public void error(TransformerException exception) throws TransformerException {
+				Log.e(TAG, exception.getMessage());
+				piped = false;
+			}
+		});
 		tHandler1.setResult(new SAXResult(tHandler2));
 		
 		XMLReader reader = XMLReaderFactory.createXMLReader();
@@ -238,7 +265,16 @@ public class BerichtsheftApp
 		
 		reader.parse(new InputSource(params));
 		
-		return true;
+		return piped;
 	}
 
+	public static String odtVorlagePath(String name) {
+		return berichtsheftPath("Vorlagen/" + name + ".odt");
+	}
+
+	public static String odtDokumentPath(String name, int...weekDate) {
+		if (isAvailable(1, weekDate))
+			name += String.format("_%d_%d", weekDate[1], weekDate[0]);
+		return berichtsheftPath("Dokumente/" + name + ".odt");
+	}
 }
